// types/reaction.enum.ts
export enum ReactionEnum {
    LIKE = 'like',
    LOVE = 'love',
    CARE = 'care',
    HAHA = 'haha',
    WOW = 'wow',
    THANK_YOU = 'thank_you',
    CLAP = 'clap',
    SUPPORT = 'support',
    INSIGHTFUL = 'insightful',
    DISLIKE = 'dislike',
    ANGRY = 'angry',
    HOAX = 'hoax',
    FAKE_NEWS = 'fake_news',
    MISLEADING = 'misleading',
    REPORT = 'report',
    DANGEROUS = 'dangerous',
    NO_REACTION = '',
}